# docs/swagger.yaml
basePath: /
definitions:
  models.Category:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  models.Product:
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
      stock:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Task Untuk Session 1.
  title: CodeWithUmam - Task Session 1
  version: "1.0"
paths:
  /api/categories:
    get:
      responses: {}
      summary: Daftar Semua Kategori
      tags:
      - categories
    post:
      parameters:
      - description: Data Kategori
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      responses: {}
      summary: Tambah Kategori
      tags:
      - categories
  /api/categories/{id}:
    delete:
      parameters:
      - description: Kategori ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: Hapus Kategori
      tags:
      - categories
    get:
      parameters:
      - description: Kategori ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
      summary: Ambil Kategori by ID
      tags:
      - categories
    put:
      parameters:
      - description: Kategori ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: Update Kategori
      tags:
      - categories
  /api/product:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
      summary: Daftar Semua Produk
      tags:
      - product
    post:
      consumes:
      - application/json
      parameters:
      - description: Data Product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Product'
      summary: Tambah Produk Baru
      tags:
      - product
  /api/product/{id}:
    delete:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: Hapus Produk
      tags:
      - product
    get:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
      summary: Ambil Produk by ID
      tags:
      - product
    put:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Data Update
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      responses: {}
      summary: Update Produk
      tags:
      - product
  /health:
    get:
      responses: {}
      summary: Health Check
      tags:
      - system
swagger: "2.0"